moved location callbacks back to individual activities; made switch selection persist
