+++ GuardiansoftheSpectrum/app/src/main/java/edu/colorado/gots/guardiansofthespectrum/BaseActivity.java
@@ -1,15 +1,13 @@
 package edu.colorado.gots.guardiansofthespectrum;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.widget.Toast;
 
-public abstract class BaseActivity extends AppCompatActivity implements LocationServicesManager.LocationServicesCallbacks {
+public abstract class BaseActivity extends AppCompatActivity {
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         MenuInflater mMenuInflater = getMenuInflater();
@@ -25,7 +23,8 @@ public abstract class BaseActivity extends AppCompatActivity implements Location
 
         switch (item.getItemId()){
             case R.id.action_scan:
-                LocationServicesManager.getInstance(getApplicationContext()).checkAndResolvePermissions(this);
+                Intent scan = new Intent(this, ScanActivity.class);
+                startActivity(scan);
                 return true;
             case R.id.action_my_info:
                 Intent info = new Intent(this, MyInfoActivity.class);
@@ -43,30 +42,5 @@ public abstract class BaseActivity extends AppCompatActivity implements Location
                 return super.onOptionsItemSelected(item);
 
         }
-    }
-
-    protected void onActivityResult(int requestCode, int returnCode, Intent i) {
-        switch (requestCode) {
-            case LocationServicesManager.LOCATION_SERVICE_RESOLUTION:
-                if (returnCode != Activity.RESULT_OK) {
-                    //changes not made successfully. just gripe for now
-                    Toast.makeText(getApplicationContext(), "Location services needed to send data", Toast.LENGTH_SHORT).show();
-                    this.onLocationNotEnabled();
-                } else {
-                    this.onLocationEnabled();
-                }
-                break;
-            default:
-                break;
-        }
-    }
-
-    public void onLocationEnabled() {
-        Intent scan = new Intent(this, ScanActivity.class);
-        startActivity(scan);
-    }
-
-    public void onLocationNotEnabled() {
-
     }
 }
