+++ GuardiansoftheSpectrum/app/src/main/java/edu/colorado/gots/guardiansofthespectrum/MainActivity.java
@@ -1,6 +1,5 @@
 package edu.colorado.gots.guardiansofthespectrum;
 
-
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
